+++ modules/log-time/init.js
